small change in 2 task

+++ src/task2.cpp
@@ -4,7 +4,7 @@
 using namespace std;
 
 bool checkPrime(unsigned int value) {
-    for (int i = 2; i <= sqrt(value); i++)
+    for (unsigned long long i = 2; i <= sqrt(value); i++)
         if (value % i == 0)
             return false;
          return true;
